+++ UI/Web/src/app/collections/all-collections/all-collections.component.ts
@@ -1,19 +1,21 @@
-import { Component, EventEmitter, OnInit } from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, EventEmitter, OnInit } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { Router } from '@angular/router';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { EditCollectionTagsComponent } from 'src/app/cards/_modals/edit-collection-tags/edit-collection-tags.component';
+import { UtilityService } from 'src/app/shared/_services/utility.service';
 import { CollectionTag } from 'src/app/_models/collection-tag';
 import { JumpKey } from 'src/app/_models/jumpbar/jump-key';
+import { Tag } from 'src/app/_models/tag';
 import { ActionItem, ActionFactoryService, Action } from 'src/app/_services/action-factory.service';
 import { CollectionTagService } from 'src/app/_services/collection-tag.service';
-import { ImageService } from 'src/app/_services/image.service';
 
 
 @Component({
   selector: 'app-all-collections',
   templateUrl: './all-collections.component.html',
-  styleUrls: ['./all-collections.component.scss']
+  styleUrls: ['./all-collections.component.scss'],
+  changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class AllCollectionsComponent implements OnInit {
 
@@ -26,7 +28,8 @@ export class AllCollectionsComponent implements OnInit {
 
   constructor(private collectionService: CollectionTagService, private router: Router,
     private actionFactoryService: ActionFactoryService, private modalService: NgbModal, 
-    private titleService: Title, private imageService: ImageService) {
+    private titleService: Title, private utilityService: UtilityService, 
+    private readonly cdRef: ChangeDetectorRef) {
     this.router.routeReuseStrategy.shouldReuseRoute = () => false;
     this.titleService.setTitle('Kavita - Collections');
   }
@@ -34,6 +37,7 @@ export class AllCollectionsComponent implements OnInit {
   ngOnInit() {
     this.loadPage();
     this.collectionTagActions = this.actionFactoryService.getCollectionTagActions(this.handleCollectionActionCallback.bind(this));
+    this.cdRef.markForCheck();
   }
 
 
@@ -43,34 +47,14 @@ export class AllCollectionsComponent implements OnInit {
   }
 
   loadPage() {
+    this.isLoading = true;
+    this.cdRef.markForCheck();
     this.collectionService.allTags().subscribe(tags => {
       this.collections = tags;
       this.isLoading = false;
-
-      const keys: {[key: string]: number} = {};
-      tags.forEach(s => {
-        let ch = s.title.charAt(0);
-        if (/\d|\#|!|%|@|\(|\)|\^|\*/g.test(ch)) {
-          ch = '#';
-        }
-        if (!keys.hasOwnProperty(ch)) {
-          keys[ch] = 0;
-        }
-        keys[ch] += 1;
-      });
-      this.jumpbarKeys = Object.keys(keys).map(k => {
-        return {
-          key: k,
-          size: keys[k],
-          title: k.toUpperCase()
-        }
-      }).sort((a, b) => {
-        if (a.key < b.key) return -1;
-        if (a.key > b.key) return 1;
-        return 0;
-      });
+      this.jumpbarKeys = this.utilityService.getJumpKeys(tags, (t: Tag) => t.title);
+      this.cdRef.markForCheck();
     });
-    
   }
 
   handleCollectionActionCallback(action: Action, collectionTag: CollectionTag) {
@@ -79,9 +63,8 @@ export class AllCollectionsComponent implements OnInit {
         const modalRef = this.modalService.open(EditCollectionTagsComponent, { size: 'lg', scrollable: true });
         modalRef.componentInstance.tag = collectionTag;
         modalRef.closed.subscribe((results: {success: boolean, coverImageUpdated: boolean}) => {
-          this.loadPage();
-          if (results.coverImageUpdated) {
-            collectionTag.coverImage = this.imageService.randomize(this.imageService.getCollectionCoverImage(collectionTag.id));
+          if (results.success) {
+            this.loadPage();
           }
         });
         break;
